+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## actions development version
 
 - `check-links` - only open an issue if there are broken links. (#105, @kelly-sovacool)
+- fix `build-docker` - use everything after the first dot in the dockerfile name as the version tag. (#109, @kelly-sovacool)
 
 ## actions 0.4.0